fix(awq): use torch half conversion helpers

+++ optimum/quanto/library/extensions/cuda/awq/v2/gemv_cuda.cu
@@ -42,7 +42,7 @@ __device__ __forceinline__ static void warp_reduce(half* psum, float (*out_smem)
       #pragma unroll
       for (int i = 0; i < Num; ++i)
       {
-          fpsum[i] = static_cast<float>(psum[i]);
+          fpsum[i] = __half2float(psum[i]);
       }
 
       #pragma unroll
@@ -97,7 +97,7 @@ __global__ void gemv_kernel(
 
     half psum[Num];
     for (int i = 0; i < Num; ++i)
-        psum[i] = static_cast<half>(0.f);
+        psum[i] = __float2half(0.f);
 
     extern __shared__ uint8_t shmem[];
     float(*out_smem)[Num * kInterleave] = reinterpret_cast<float(*)[Num * kInterleave]>(shmem);
@@ -199,7 +199,7 @@ __global__ void gemv_kernel(
         {
             acc += out_smem[j][i];
         }
-        outputs[batch_idx * OC + blk_row_offset + oc_idx] = static_cast<half>(acc);
+        outputs[batch_idx * OC + blk_row_offset + oc_idx] = __float2half(acc);
     }
 }
